changed int to long

+++ src/main/java/com/bae/accountmaker/service/AccountNumberGenerator.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 @Service
 public interface AccountNumberGenerator {
 
-	Integer generateNumber();
+	Long generateNumber();
 
 	Character generateCharacter();
